Also display build number in app settings
